Extract duplicate values and logic to common module
