+++ fabric/src/main/java/de/guntram/mcmod/antighost/fabric/AntiGhost.java
@@ -1,67 +1,34 @@
 package de.guntram.mcmod.antighost.fabric;
 
+import de.guntram.mcmod.antighost.BlockRefresher;
 import de.guntram.mcmod.crowdintranslate.CrowdinTranslate;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.network.ClientPlayNetworkHandler;
-import net.minecraft.client.network.ClientPlayerEntity;
-import net.minecraft.client.option.KeyBinding;
-import net.minecraft.network.packet.c2s.play.PlayerActionC2SPacket;
-import net.minecraft.text.Text;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
-import static org.lwjgl.glfw.GLFW.GLFW_KEY_G;
+
+import static de.guntram.mcmod.antighost.Constants.*;
 
 public class AntiGhost implements ClientModInitializer
 {
-    static final String MODID="antighost";
-    static KeyBinding requestBlocks;
-    
     @Override
     public void onInitializeClient()
     {
-        final String category="key.categories.antighost";
-        requestBlocks = new KeyBinding("key.antighost.reveal", GLFW_KEY_G, category);
-        CrowdinTranslate.downloadTranslations(MODID);
-        KeyBindingHelper.registerKeyBinding(requestBlocks);
+        CrowdinTranslate.downloadTranslations(MOD_ID);
+        KeyBindingHelper.registerKeyBinding(REFRESH_KEY_BINDING);
         ClientTickEvents.END_CLIENT_TICK.register(e->keyPressed());
-        ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> {
-            dispatcher.register(
-                ClientCommandManager.literal("ghost").executes(c -> {
-                    this.execute();
-                    return 0;
-                })
-            );
-        });
+        ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> dispatcher.register(
+            ClientCommandManager.literal(REFRESH_COMMAND_NAME).executes( c-> {
+                BlockRefresher.refreshBlocks();
+                return 0;
+            })
+        ));
     }
 
     public void keyPressed() {
-        ClientPlayerEntity player = MinecraftClient.getInstance().player;
-        if (requestBlocks.wasPressed()) {
-            this.execute();
-            player.sendMessage(Text.translatable("msg.request"), false);
+        if (REFRESH_KEY_BINDING.wasPressed()) {
+            BlockRefresher.refreshBlocks();
         }
-    }
-    
-    public void execute() {
-        MinecraftClient mc=MinecraftClient.getInstance();
-        ClientPlayNetworkHandler conn = mc.getNetworkHandler();
-        if (conn==null)
-            return;
-        BlockPos pos=mc.player.getBlockPos();
-        for (int dx=-4; dx<=4; dx++)
-            for (int dy=-4; dy<=4; dy++)
-                for (int dz=-4; dz<=4; dz++) {
-                    PlayerActionC2SPacket packet=new PlayerActionC2SPacket(
-                            PlayerActionC2SPacket.Action.ABORT_DESTROY_BLOCK, 
-                            new BlockPos(pos.getX()+dx, pos.getY()+dy, pos.getZ()+dz),
-                            Direction.UP       // with ABORT_DESTROY_BLOCK, this value is unused
-                    );
-                    conn.sendPacket(packet);
-                }
     }
 }
